update to include CRUD placeholder from DB team, and import Weather_obj from model.py
