+++ 7-finish-and-enhance-db-and-api/code_skeletons/api_team/model.py
@@ -1,0 +1,9 @@
+class Weather_obj:
+
+    def __init__(self, time, summary, temperature, precipitation_intensity, precipitation_type, precipitation_probability):
+            self.time = time
+            self.summary = summary
+            self.temperature = temperature
+            self.precipitation_intensity = precipitation_intensity
+            self.precipitation_type = precipitation_type
+            self.precipitation_probability = precipitation_probability
